Agent Evacuation to Tennis Courts
I've made changes to the AgentController script so agents can calculate their destination coordinates even if the evacuation boundary is rotated. The agents are able to form an orderly grid inside the tennis courts however, I believe these calculations are more computationally expensive. This means that in a large simulation with 1200 agents, as they currently pathfind sequentially, the simulation is not accurate as the congestion does not occur.