Agent Spawing Confined to School Building
I've added multiple spawning boundaries to unity for the agent to randomly select from. As each of the spawning boundaries are different sizes, the code gives each boundary a 'weight' based on its area. The random boundary selector then chooses a random boundary using this weighting system to make the agents spawn evenly throughout the building